Changed naming structure (read below)
Following circleci's (and other's) lead, I'm changing the folder
structure to correspond to the tag names:

    soutech/<language>:<version>-<osversion>[-<layer>]
    soutech/ruby:2.6-stretch-node-awscli

`ruby` is now the base name of the image, with each version of ruby
being published as a tag. This will also make it easier to write an
automated script to write tags and upload to dockerhub.