Add Ruby 2.7.6 Docker image for deployment
This creates a 2.7.6 version of the Docker
image being used for the CircleCI deployment
process.

In the 2.6 version, the 2.6-stretch-node circleci
variant was used, but there is no equivalent in the
cimg repo (the new CircleCI repo). Since the same
image is being used for running the test suite as
for deploying, we are choosing to use the -browsers
variant here which has Node.js, Selenium, and browser
dependencies pre-installed.