Dockerfile updates (#4)
This adds two new Dockerfiles to be used in the deploy
flow for prosecco. The 2.7.6-browsers file which uses a 
CircleCI image did not work since it used NodeJS version 
16, and we require version 14. This file was removed. A new 
Dockerfile, 2.7.6 was created that uses a Ruby Docker image
as the base, on which we install NodeJS 14. This image is 
used to run the specs, and precompile the assets.

A new 2.7.6 Alpine Linux Docker image that is used as a base 
for the final deploy image was also added. As compared to the 
2.6 version, I removed yarn, as the assets are already precompiled 
at the stage where this is used. Also, the version of Node does not
 matter at this stage either since the assets are already precompiled. 
I also removed installing the bundler gem and made that the
 responsibility of the Dockerfile in prosecco.